+++ tests/test_pytorch/test_precision.py
@@ -28,6 +28,7 @@ elif module_available("pytorch_lightning"):
     from pytorch_lightning.plugins import MixedPrecisionPlugin
 
 from lightning_habana.pytorch.accelerator import HPUAccelerator
+from lightning_habana.pytorch.plugins.precision import HPUPrecisionPlugin
 from lightning_habana.pytorch.strategies.single import SingleHPUStrategy
 
 
@@ -35,6 +36,12 @@ from lightning_habana.pytorch.strategies.single import SingleHPUStrategy
 def does_not_raise():
     """No-op context manager as a complement to pytest.raises."""
     yield
+
+
+@pytest.fixture()
+def hmp_params(request):
+    """Returns params for HPUPrecisionPlugin."""
+    return {"device": "hpu"}
 
 
 @pytest.fixture()
@@ -117,6 +124,7 @@ class BMAutocastDecorator(BaseBMActive):
 @pytest.mark.parametrize(
     ("plugin", "params"),
     [
+        (HPUPrecisionPlugin, "hmp_params"),
         (MixedPrecisionPlugin, "mpp_params"),
     ],
 )
@@ -129,6 +137,7 @@ def test_precision_plugins_instance(plugin, params, request):
 @pytest.mark.parametrize(
     ("plugin", "params"),
     [
+        (HPUPrecisionPlugin, "hmp_params"),
         (MixedPrecisionPlugin, "mpp_params"),
     ],
 )
@@ -157,17 +166,30 @@ def test_mixed_precision_plugin(tmpdir, plugin, params, request):
         trainer.fit(model)
 
 
+def test_unsupported_precision_plugin():
+    """Tests unsupported HPUPrecisionPlugin init."""
+    with pytest.raises(ValueError, match=r"accelerator='hpu', precision='mixed'\)` is not supported."):
+        HPUPrecisionPlugin(precision="mixed")
+
+
 @pytest.mark.parametrize(
     ("model", "plugin", "params", "expectation"),
     [
         (BMAutocastCM, [], "", does_not_raise()),
         (BMAutocastDecorator, [], "", does_not_raise()),
         (BaseBMActive, MixedPrecisionPlugin, "mpp_params", does_not_raise()),
+        (
+            BaseBMActive,
+            HPUPrecisionPlugin,
+            "hmp_params",
+            does_not_raise(),
+        ),
     ],
     ids=[
         "TorchAutocast_CM_True",
         "TorchAutocast_Decorator_True",
         "MixedPrecisionPlugin_True",
+        "HPUPrecisionPlugin_False",
     ],
 )
 def test_mixed_precision_autocast_active(tmpdir, model, plugin, params, expectation, request):
@@ -185,13 +207,29 @@ def test_mixed_precision_autocast_active(tmpdir, model, plugin, params, expectat
     "model_plugin_list",
     [
         [
+            (BaseBM, HPUPrecisionPlugin, "hmp_params"),
+            (BMAutocastCM, [], ""),
+        ],
+        [
+            (BaseBM, HPUPrecisionPlugin, "hmp_params"),
+            (BMAutocastDecorator, [], ""),
+        ],
+        [
+            (BaseBM, HPUPrecisionPlugin, "hmp_params"),
+            (BaseBM, MixedPrecisionPlugin, "mpp_params"),
+        ],
+        [
+            (BaseBM, HPUPrecisionPlugin, "hmp_params"),
             (BMAutocastCM, [], ""),
             (BMAutocastDecorator, [], ""),
             (BaseBM, MixedPrecisionPlugin, "mpp_params"),
         ],
     ],
     ids=[
-        "AutocastCM_AutocastDecorator_MixedPrecisionPlugin",
+        "HPUPrecisionPlugin_AutocastCM",
+        "HPUPrecisionPlugin_AutocastDecorator",
+        "HPUPrecisionPlugin_MixedPrecisionPlugin",
+        "HPUPrecisionPlugin_AutocastCM_AutocastDecorator_MixedPrecisionPlugin",
     ],
 )
 def test_mixed_precision_compare_accuracy(tmpdir, model_plugin_list, request):